+++ app/src/main/java/com/nikola/jakshic/dagger/matchstats/Models.kt
@@ -10,16 +10,16 @@ import com.squareup.moshi.JsonClass
 class MatchStatsJson {
     @Json(name = "match_id") var matchId: Long = 0
     @Json(name = "radiant_win") var isRadiantWin = false
-    @Json(name = "dire_score") var direScore: Int = 0
-    @Json(name = "radiant_score") var radiantScore: Int = 0
-    @Json(name = "skill") var skill: Int = 0
-    @Json(name = "game_mode") var mode: Int = 0
-    @Json(name = "duration") var duration: Int = 0
+    @Json(name = "dire_score") var direScore: Long = 0
+    @Json(name = "radiant_score") var radiantScore: Long = 0
+    @Json(name = "skill") var skill: Long = 0
+    @Json(name = "game_mode") var mode: Long = 0
+    @Json(name = "duration") var duration: Long = 0
     @Json(name = "start_time") var startTime: Long = 0
-    @Json(name = "barracks_status_radiant") var radiantBarracks: Int = 0
-    @Json(name = "barracks_status_dire") var direBarracks: Int = 0
-    @Json(name = "tower_status_radiant") var radiantTowers: Int = 0
-    @Json(name = "tower_status_dire") var direTowers: Int = 0
+    @Json(name = "barracks_status_radiant") var radiantBarracks: Long = 0
+    @Json(name = "barracks_status_dire") var direBarracks: Long = 0
+    @Json(name = "tower_status_radiant") var radiantTowers: Long = 0
+    @Json(name = "tower_status_dire") var direTowers: Long = 0
     @Json(name = "radiant_team") var radiantTeam: Team? = null
     @Json(name = "dire_team") var direTeam: Team? = null
     @Json(name = "players") var players: List<PlayerStatsJson>? = null
@@ -34,31 +34,31 @@ data class PlayerStatsJson(
     @Json(name = "match_id") var matchId: Long,
     @Json(name = "name") var name: String?,
     @Json(name = "personaname") var personaName: String?,
-    @Json(name = "player_slot") var playerSlot: Int,
-    @Json(name = "assists") var assists: Int,
-    @Json(name = "backpack_0") var backpack0: Int,
-    @Json(name = "backpack_1") var backpack1: Int,
-    @Json(name = "backpack_2") var backpack2: Int,
-    @Json(name = "deaths") var deaths: Int,
-    @Json(name = "denies") var denies: Int,
-    @Json(name = "gold_per_min") var goldPerMin: Int,
-    @Json(name = "hero_damage") var heroDamage: Int,
-    @Json(name = "hero_healing") var heroHealing: Int,
-    @Json(name = "hero_id") var heroId: Int,
-    @Json(name = "item_0") var item0: Int,
-    @Json(name = "item_1") var item1: Int,
-    @Json(name = "item_2") var item2: Int,
-    @Json(name = "item_3") var item3: Int,
-    @Json(name = "item_4") var item4: Int,
-    @Json(name = "item_5") var item5: Int,
-    @Json(name = "item_neutral") var itemNeutral: Int,
-    @Json(name = "kills") var kills: Int,
-    @Json(name = "last_hits") var lastHits: Int,
-    @Json(name = "level") var level: Int,
-    @Json(name = "tower_damage") var towerDamage: Int,
-    @Json(name = "xp_per_min") var xpPerMin: Int,
-    @Json(name = "purchase_ward_observer") var purchaseWardObserver: Int = 0,
-    @Json(name = "purchase_ward_sentry") var purchaseWardSentry: Int = 0
+    @Json(name = "player_slot") var playerSlot: Long,
+    @Json(name = "assists") var assists: Long,
+    @Json(name = "backpack_0") var backpack0: Long,
+    @Json(name = "backpack_1") var backpack1: Long,
+    @Json(name = "backpack_2") var backpack2: Long,
+    @Json(name = "deaths") var deaths: Long,
+    @Json(name = "denies") var denies: Long,
+    @Json(name = "gold_per_min") var goldPerMin: Long,
+    @Json(name = "hero_damage") var heroDamage: Long,
+    @Json(name = "hero_healing") var heroHealing: Long,
+    @Json(name = "hero_id") var heroId: Long,
+    @Json(name = "item_0") var item0: Long,
+    @Json(name = "item_1") var item1: Long,
+    @Json(name = "item_2") var item2: Long,
+    @Json(name = "item_3") var item3: Long,
+    @Json(name = "item_4") var item4: Long,
+    @Json(name = "item_5") var item5: Long,
+    @Json(name = "item_neutral") var itemNeutral: Long,
+    @Json(name = "kills") var kills: Long,
+    @Json(name = "last_hits") var lastHits: Long,
+    @Json(name = "level") var level: Long,
+    @Json(name = "tower_damage") var towerDamage: Long,
+    @Json(name = "xp_per_min") var xpPerMin: Long,
+    @Json(name = "purchase_ward_observer") var purchaseWardObserver: Long = 0,
+    @Json(name = "purchase_ward_sentry") var purchaseWardSentry: Long = 0
 )
 
 data class MatchStatsUI(
@@ -114,16 +114,16 @@ fun MatchStatsJson.mapToDb(): Match_stats {
     return Match_stats(
         match_id = matchId,
         radiant_win = isRadiantWin,
-        dire_score = direScore.toLong(),
-        radiant_score = radiantScore.toLong(),
-        skill = skill.toLong(),
-        game_mode = mode.toLong(),
-        duration = duration.toLong(),
+        dire_score = direScore,
+        radiant_score = radiantScore,
+        skill = skill,
+        game_mode = mode,
+        duration = duration,
         start_time = startTime,
-        radiant_barracks = radiantBarracks.toLong(),
-        dire_barracks = direBarracks.toLong(),
-        radiant_towers = radiantTowers.toLong(),
-        dire_towers = direTowers.toLong(),
+        radiant_barracks = radiantBarracks,
+        dire_barracks = direBarracks,
+        radiant_towers = radiantTowers,
+        dire_towers = direTowers,
         radiant_name = radiantTeam?.name,
         dire_name = direTeam?.name
     )
@@ -192,30 +192,30 @@ fun PlayerStatsJson.mapToDb(): Player_stats {
         match_id = matchId,
         name = name,
         persona_name = personaName,
-        player_slot = playerSlot.toLong(),
-        assists = assists.toLong(),
-        backpack_0 = backpack0.toLong(),
-        backpack_1 = backpack1.toLong(),
-        backpack_2 = backpack2.toLong(),
-        deaths = deaths.toLong(),
-        denies = denies.toLong(),
-        gpm = goldPerMin.toLong(),
-        hero_damage = heroDamage.toLong(),
-        hero_healing = heroHealing.toLong(),
-        hero_id = heroId.toLong(),
-        item_0 = item0.toLong(),
-        item_1 = item1.toLong(),
-        item_2 = item2.toLong(),
-        item_3 = item3.toLong(),
-        item_4 = item4.toLong(),
-        item_5 = item5.toLong(),
-        item_neutral = itemNeutral.toLong(),
-        kills = kills.toLong(),
-        last_hits = lastHits.toLong(),
-        level = level.toLong(),
-        tower_damage = towerDamage.toLong(),
-        xpm = xpPerMin.toLong(),
-        observers = purchaseWardObserver.toLong(),
-        sentries = purchaseWardSentry.toLong()
+        player_slot = playerSlot,
+        assists = assists,
+        backpack_0 = backpack0,
+        backpack_1 = backpack1,
+        backpack_2 = backpack2,
+        deaths = deaths,
+        denies = denies,
+        gpm = goldPerMin,
+        hero_damage = heroDamage,
+        hero_healing = heroHealing,
+        hero_id = heroId,
+        item_0 = item0,
+        item_1 = item1,
+        item_2 = item2,
+        item_3 = item3,
+        item_4 = item4,
+        item_5 = item5,
+        item_neutral = itemNeutral,
+        kills = kills,
+        last_hits = lastHits,
+        level = level,
+        tower_damage = towerDamage,
+        xpm = xpPerMin,
+        observers = purchaseWardObserver,
+        sentries = purchaseWardSentry
     )
 }
