+++ app/src/main/java/com/nikola/jakshic/dagger/profile/heroes/HeroAdapter.kt
@@ -9,7 +9,7 @@ import com.nikola.jakshic.dagger.common.inflate
 import com.nikola.jakshic.dagger.util.DotaUtil
 import kotlinx.android.synthetic.main.item_hero.view.*
 
-class HeroAdapter(val listener: (Int) -> Unit) : RecyclerView.Adapter<HeroAdapter.HeroVH>() {
+class HeroAdapter(val listener: (Long) -> Unit) : RecyclerView.Adapter<HeroAdapter.HeroVH>() {
 
     private var list: List<HeroUI>? = null
 
@@ -36,9 +36,9 @@ class HeroAdapter(val listener: (Int) -> Unit) : RecyclerView.Adapter<HeroAdapte
 
         fun bind(item: HeroUI) {
             with(itemView) {
-                imgHero.load(DotaUtil.getHero(context, item.heroId.toLong()))
+                imgHero.load(DotaUtil.getHero(context, item.heroId))
                 tvGamesPlayed.text = item.gamesPlayed.toString()
-                val winRate = if (item.gamesPlayed != 0) (item.gamesWon.toFloat() / item.gamesPlayed) * 100 else 0f
+                val winRate = if (item.gamesPlayed != 0L) (item.gamesWon.toFloat() / item.gamesPlayed) * 100 else 0f
                 tvHeroWinRate.text = context.resources.getString(R.string.hero_winrate, winRate)
                 setPercentageBar(viewHeroWinRateBar, winRate)
                 tvHeroWinLose.text = context.resources.getString(R.string.hero_win_loss,
