+++ app/src/main/java/com/nikola/jakshic/dagger/profile/matches/Models.kt
@@ -8,24 +8,24 @@ import com.squareup.moshi.JsonClass
 data class MatchJson(
     var accountId: Long = 0,
     @Json(name = "match_id") var matchId: Long,
-    @Json(name = "hero_id") var heroId: Int,
-    @Json(name = "player_slot") var playerSlot: Int,
-    @Json(name = "skill") var skill: Int,
-    @Json(name = "duration") var duration: Int,
-    @Json(name = "game_mode") var gameMode: Int,
-    @Json(name = "lobby_type") var lobbyType: Int,
+    @Json(name = "hero_id") var heroId: Long,
+    @Json(name = "player_slot") var playerSlot: Long,
+    @Json(name = "skill") var skill: Long,
+    @Json(name = "duration") var duration: Long,
+    @Json(name = "game_mode") var gameMode: Long,
+    @Json(name = "lobby_type") var lobbyType: Long,
     @Json(name = "radiant_win") var isRadiantWin: Boolean,
     @Json(name = "start_time") var startTime: Long
 )
 
 data class MatchUI(
     val matchId: Long,
-    val heroId: Int,
-    val playerSlot: Int,
-    val skill: Int,
-    val duration: Int,
-    val gameMode: Int,
-    val lobbyType: Int,
+    val heroId: Long,
+    val playerSlot: Long,
+    val skill: Long,
+    val duration: Long,
+    val gameMode: Long,
+    val lobbyType: Long,
     val isRadiantWin: Boolean,
     val startTime: Long
 )
@@ -48,12 +48,12 @@ fun MatchJson.mapToDb(): Matches {
     return Matches(
         account_id = this.accountId,
         match_id = this.matchId,
-        hero_id = this.heroId.toLong(),
-        player_slot = this.playerSlot.toLong(),
-        skill = this.skill.toLong(),
-        duration = this.duration.toLong(),
-        mode = this.gameMode.toLong(),
-        lobby = this.lobbyType.toLong(),
+        hero_id = this.heroId,
+        player_slot = this.playerSlot,
+        skill = this.skill,
+        duration = this.duration,
+        mode = this.gameMode,
+        lobby = this.lobbyType,
         radiant_win = this.isRadiantWin,
         start_time = this.startTime
     )
@@ -62,12 +62,12 @@ fun MatchJson.mapToDb(): Matches {
 fun Matches.mapToUi(): MatchUI {
     return MatchUI(
         matchId = this.match_id,
-        heroId = this.hero_id.toInt(),
-        playerSlot = this.player_slot.toInt(),
-        skill = this.skill.toInt(),
-        duration = this.duration.toInt(),
-        gameMode = this.mode.toInt(),
-        lobbyType = this.lobby.toInt(),
+        heroId = this.hero_id,
+        playerSlot = this.player_slot,
+        skill = this.skill,
+        duration = this.duration,
+        gameMode = this.mode,
+        lobbyType = this.lobby,
         isRadiantWin = this.radiant_win,
         startTime = this.start_time
     )
@@ -87,12 +87,12 @@ fun mapToUi(
 ): MatchUI {
     return MatchUI(
         matchId = match_id,
-        heroId = hero_id.toInt(),
-        playerSlot = player_slot.toInt(),
-        skill = skill.toInt(),
-        duration = duration.toInt(),
-        gameMode = mode.toInt(),
-        lobbyType = lobby.toInt(),
+        heroId = hero_id,
+        playerSlot = player_slot,
+        skill = skill,
+        duration = duration,
+        gameMode = mode,
+        lobbyType = lobby,
         isRadiantWin = radiant_win,
         startTime = start_time
     )
